Refactored the TODOs for GLSLCompiler-only functionality. Adds a file-backed ICPUBuffer to IFile
