+++ include/nbl/asset/utils/CGLSLCompiler.h
@@ -129,6 +129,15 @@ class NBL_API2 CGLSLCompiler final : public IShaderCompiler
 
 		std::string preprocessShader(std::string&& code, IShader::E_SHADER_STAGE& stage, const SPreprocessorOptions& preprocessOptions, std::vector<CCache::SEntry::SPreprocessingDependency>* dependencies = nullptr) const override;
 
+		static std::string escapeFilename(std::string&& code);
+
+		static void disableAllDirectivesExceptIncludes(std::string& _code);
+
+		static void reenableDirectives(std::string& _code);
+
+		static std::string encloseWithinExtraInclGuards(std::string&& _code, uint32_t _maxInclusions, const char* _identifier);
+
+		static uint32_t encloseWithinExtraInclGuardsLeadingLines(uint32_t _maxInclusions);
 	protected:
 
 		void insertIntoStart(std::string& code, std::ostringstream&& ins) const override;
