+++ include/nbl/system/IFile.h
@@ -2,6 +2,7 @@
 #define _NBL_SYSTEM_I_FILE_H_INCLUDED_
 
 #include "nbl/system/ISystem.h"
+#include "nbl/asset/ICPUBuffer.h"
 
 namespace nbl::system
 {
@@ -105,6 +106,19 @@ class IFile : public IFileBase, private ISystem::IFutureManipulator
 		}
 };
 
+}
+
+namespace nbl::asset {
+	class IFileBackedCPUBuffer : public ICPUBuffer {
+
+		IFileBackedCPUBuffer(core::smart_refctd_ptr<nbl::system::IFile> _file)
+			: ICPUBuffer(_file->getSize(), file->getMappedPointer()), file(_file) {}
+
+
+	private:
+		core::smart_refctd_ptr<nbl::system::IFile> file;
+
+	};
 }
 
 #endif
