+++ src/nbl/asset/utils/CGLSLCompiler.cpp
@@ -18,6 +18,9 @@ static constexpr const char* PREPROC_GL__DISABLER = "_this_is_a_GL__prefix_";
 static constexpr const char* PREPROC_GL__ENABLER = PREPROC_GL__DISABLER;
 static constexpr const char* PREPROC_LINE_CONTINUATION_DISABLER = "_this_is_a_line_continuation_\n";
 static constexpr const char* PREPROC_LINE_CONTINUATION_ENABLER = "_this_is_a_line_continuation_";
+//string to be replaced with all "#" except those in "#include"
+static constexpr const char* PREPROC_DIRECTIVE_DISABLER = "_this_is_a_hash_";
+static constexpr const char* PREPROC_DIRECTIVE_ENABLER = PREPROC_DIRECTIVE_DISABLER;
 
 static void disableGlDirectives(std::string& _code)
 {
@@ -100,9 +103,9 @@ namespace nbl::asset::impl
             {
                 auto res_str = std::move(result.contents);
                 //employ encloseWithinExtraInclGuards() in order to prevent infinite loop of (not necesarilly direct) self-inclusions while other # directives (incl guards among them) are disabled
-                IShaderCompiler::disableAllDirectivesExceptIncludes(res_str);
+                CGLSLCompiler::disableAllDirectivesExceptIncludes(res_str);
                 disableGlDirectives(res_str);
-                res_str = IShaderCompiler::encloseWithinExtraInclGuards(std::move(res_str), m_maxInclCnt, name.string().c_str());
+                res_str = CGLSLCompiler::encloseWithinExtraInclGuards(std::move(res_str), m_maxInclCnt, name.string().c_str());
 
                 res->content_length = res_str.size();
                 res->content = new char[res_str.size() + 1u];
@@ -142,7 +145,7 @@ std::string CGLSLCompiler::preprocessShader(std::string&& code, IShader::E_SHADE
             insertion << "#define " << define.identifier << " " << define.definition << "\n";
         insertIntoStart(code,std::move(insertion));
     }
-    IShaderCompiler::disableAllDirectivesExceptIncludes(code);
+    disableAllDirectivesExceptIncludes(code);
     disableGlDirectives(code);
     shaderc::Compiler comp;
     shaderc::CompileOptions options;
@@ -161,11 +164,94 @@ std::string CGLSLCompiler::preprocessShader(std::string&& code, IShader::E_SHADE
     }
 
     auto resolvedString = std::string(res.cbegin(), std::distance(res.cbegin(), res.cend()));
-    IShaderCompiler::reenableDirectives(resolvedString);
+    reenableDirectives(resolvedString);
     reenableGlDirectives(resolvedString);
     return resolvedString;
 }
 
+std::string CGLSLCompiler::escapeFilename(std::string&& code)
+{
+    std::string dest;
+    dest.reserve(code.size() * 2);
+    for (char c : code)
+    {
+        if (c == '\\')
+            dest.append("\\" "\\");
+        else
+            dest.push_back(c);
+    }
+    return dest;
+}
+
+//all "#", except those in "#include"/"#version"/"#pragma shader_stage(...)", replaced with `PREPROC_DIRECTIVE_DISABLER`
+void CGLSLCompiler::disableAllDirectivesExceptIncludes(std::string& _code)
+{
+    // TODO: replace this with a proper-ish proprocessor and includer one day
+    std::regex directive("#(?!(( |\t|\r|\v|\f)*(include|version|pragma shader_stage)))");//all # not followed by "include" nor "version" nor "pragma shader_stage"
+    //`#pragma shader_stage(...)` is needed for determining shader stage when `_stage` param of IShaderCompiler functions is set to ESS_UNKNOWN
+    _code = std::regex_replace(_code, directive, PREPROC_DIRECTIVE_DISABLER);
+}
+
+void CGLSLCompiler::reenableDirectives(std::string& _code)
+{
+    std::regex directive(PREPROC_DIRECTIVE_ENABLER);
+    _code = std::regex_replace(_code, directive, "#");
+}
+
+std::string CGLSLCompiler::encloseWithinExtraInclGuards(std::string&& _code, uint32_t _maxInclusions, const char* _identifier)
+{
+    assert(_maxInclusions != 0u);
+
+    using namespace std::string_literals;
+    std::string defBase_ = "_GENERATED_INCLUDE_GUARD_"s + _identifier + "_";
+    std::replace_if(defBase_.begin(), defBase_.end(), [](char c) ->bool { return !::isalpha(c) && !::isdigit(c); }, '_');
+
+    auto genDefs = [&defBase_, _maxInclusions, _identifier] {
+        auto defBase = [&defBase_](uint32_t n) { return defBase_ + std::to_string(n); };
+        std::string defs = "#ifndef " + defBase(0) + "\n\t#define " + defBase(0) + "\n";
+        for (uint32_t i = 1u; i <= _maxInclusions; ++i) {
+            const std::string defname = defBase(i);
+            defs += "#elif !defined(" + defname + ")\n\t#define " + defname + "\n";
+        }
+        defs += "#endif\n";
+        return defs;
+        };
+    auto genUndefs = [&defBase_, _maxInclusions, _identifier] {
+        auto defBase = [&defBase_](int32_t n) { return defBase_ + std::to_string(n); };
+        std::string undefs = "#ifdef " + defBase(_maxInclusions) + "\n\t#undef " + defBase(_maxInclusions) + "\n";
+        for (int32_t i = _maxInclusions - 1; i >= 0; --i) {
+            const std::string defname = defBase(i);
+            undefs += "#elif defined(" + defname + ")\n\t#undef " + defname + "\n";
+        }
+        undefs += "#endif\n";
+        return undefs;
+        };
+
+    std::string identifier = escapeFilename(_identifier);
+    return
+        genDefs() +
+        "\n"
+        "#ifndef " + defBase_ + std::to_string(_maxInclusions) +
+        "\n" +
+        // This will get turned back into #line after the directives get re-enabled
+        PREPROC_DIRECTIVE_DISABLER + "line 1 \"" + identifier.c_str() + "\"\n" +
+        _code +
+        "\n"
+        "#endif"
+        "\n\n" +
+        genUndefs();
+}
+
+// Amount of lines before the #line after having run encloseWithinExtraInclGuards
+uint32_t CGLSLCompiler::encloseWithinExtraInclGuardsLeadingLines(uint32_t _maxInclusions)
+{
+    auto lineDirectiveString = std::string(PREPROC_DIRECTIVE_DISABLER) + "line";
+    std::string str = encloseWithinExtraInclGuards(std::string(""), _maxInclusions, "encloseWithinExtraInclGuardsLeadingLines");
+    size_t lineDirectivePos = str.find(lineDirectiveString);
+    auto substr = str.substr(0, lineDirectivePos - lineDirectiveString.length());
+
+    return std::count(substr.begin(), substr.end(), '\n');
+}
 core::smart_refctd_ptr<ICPUShader> CGLSLCompiler::compileToSPIRV_impl(const std::string_view code, const IShaderCompiler::SCompilerOptions& options, std::vector<CCache::SEntry::SPreprocessingDependency>* dependencies) const
 {
     // The dependencies are only sent if a Cache was requested. Since caching is not supported for GLSL, we crash the program
